requirements.txt

torch
Pillow
matplotlib
tensorflow
streamlit
transformers
